+++ src/main/java/org/tinyconfiguration/imp/basic/io/HandlerJSON.java
@@ -366,7 +366,6 @@ public final class HandlerJSON {
          * @param instance The configuration instance to read
          * @return Future object representing the reading task
          * @throws CompletionException If any exceptions occurs at runtime
-         * @see HandlerJSON#READER#read(Configuration)
          */
         @Override
         public Future<Void> readAsync(Configuration instance) {